+++ example-crates/external-start/src/main.rs
@@ -77,5 +77,5 @@ unsafe fn origin_main(_argc: usize, _argv: *mut *mut u8, _envp: *mut *mut u8) ->
 #[no_mangle]
 unsafe fn main(_argc: i32, _argv: *mut *mut u8, _envp: *mut *mut u8) -> i32 {
     eprintln!("Main was not supposed to be called!");
-    program::abort();
+    program::trap();
 }