+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# origin 0.23
+
+## Changes
+
+`program::abort` has been renamed to `program::trap`, and the
+"panic-handler-abort" feature has been renamed to "panic-handler-trap".
+
 # origin 0.22
 
 ## Changes